renamed obj folder to models
