+++ tests/unittests/parser.py
@@ -8,16 +8,16 @@ sys.path.append("..")
 from vvspy.trip import _parse_response as parse_trip_response
 from vvspy.arrivals import _parse_response as parse_arrival_response
 from vvspy.departures import _parse_response as parse_departure_response
-from vvspy.obj import Trip
-from vvspy.obj import Arrival
-from vvspy.obj import Departure
+from vvspy.models import Trip
+from vvspy.models import Arrival
+from vvspy.models import Departure
 
 path = os.path.join(".", "mock_results")
 
 
 class TestParser(unittest.TestCase):
     def trip_200(self):
-        with open(os.path.join(path, "trip_request_200.json"), 'r') as f:
+        with open(os.path.join(path, "trip_request_200.json"), "r") as f:
             data = json.load(f)
         results = parse_trip_response(data)
         self.assertTrue(results)
@@ -25,13 +25,13 @@ class TestParser(unittest.TestCase):
         self.assertEqual(results[0].connections[0].origin.delay, 1)
 
     def trip_400(self):
-        with open(os.path.join(path, "trip_request_400.json"), 'r') as f:
+        with open(os.path.join(path, "trip_request_400.json"), "r") as f:
             data = json.load(f)
         results = parse_trip_response(data)
         self.assertEqual(results, [])
 
     def arrival_200(self):
-        with open(os.path.join(path, "arrival_request_200.json"), 'r') as f:
+        with open(os.path.join(path, "arrival_request_200.json"), "r") as f:
             data = json.load(f)
         results = parse_arrival_response(data)
         self.assertTrue(results)
@@ -39,13 +39,13 @@ class TestParser(unittest.TestCase):
         self.assertEqual(results[0].delay, 1)
 
     def arrival_400(self):
-        with open(os.path.join(path, "arrival_request_400.json"), 'r') as f:
+        with open(os.path.join(path, "arrival_request_400.json"), "r") as f:
             data = json.load(f)
         results = parse_arrival_response(data)
         self.assertEqual(results, [])
 
     def departure_200(self):
-        with open(os.path.join(path, "departure_request_200.json"), 'r') as f:
+        with open(os.path.join(path, "departure_request_200.json"), "r") as f:
             data = json.load(f)
         results = parse_departure_response(data)
         self.assertTrue(results)
@@ -53,7 +53,7 @@ class TestParser(unittest.TestCase):
         self.assertEqual(results[0].delay, 1)
 
     def departure_400(self):
-        with open(os.path.join(path, "departure_request_400.json"), 'r') as f:
+        with open(os.path.join(path, "departure_request_400.json"), "r") as f:
             data = json.load(f)
         results = parse_departure_response(data)
         self.assertEqual(results, [])
