+++ vvspy/models/arrival.py
