+++ vvspy/models/connection.py
