+++ vvspy/models/departure.py
