+++ vvspy/models/line_operator.py
