+++ vvspy/models/trip.py
